+++ package/src/types/types.ts
@@ -1,5 +1,7 @@
 import type { ExtendableGenerics, LiteralStringForUnion } from 'stream-chat';
 
+import type { FileStateValue } from '../utils/utils';
+
 export type Asset = {
   duration: number;
   height: number;
@@ -20,6 +22,26 @@ export type File = {
   size?: number;
   // The uri should be of type `string`. But is `string|undefined` because the same type is used for the response from Stream's Attachment. This shall be fixed.
   uri?: string;
+};
+
+export type FileUpload = {
+  file: File;
+  id: string;
+  state: FileStateValue;
+  duration?: number;
+  paused?: boolean;
+  progress?: number;
+  thumb_url?: string;
+  url?: string;
+};
+
+export type ImageUpload = {
+  file: Partial<Asset>;
+  id: string;
+  state: FileStateValue;
+  height?: number;
+  url?: string;
+  width?: number;
 };
 
 export type DefaultAttachmentType = UnknownType & {
